theme script migration
