+++ Turntable/README.md
@@ -43,19 +43,18 @@ spicetify config extensions fullAppDisplay.js
 spicetify apply
 ```
 
-2. put **Turntable** and **turntable.js** into the **.config/spicetify**
+1. put **Turntable** into the **.config/spicetify**
 
 ```shell
 cd spicetify-themes
 cp -r Turntable ~/.config/spicetify/Themes
-cp Turntable/turntable.js ~/.config/spicetify/Extensions
 ```
 
-3. select the theme and extension, then apply
+1. select the theme, then apply
 
 ```shell
 spicetify config current_theme Turntable
-spicetify config extensions turntable.js
+spicetify config inject_theme_js 1
 spicetify apply
 ```
 
@@ -65,7 +64,6 @@ spicetify apply
 
 ```shell
 rm -r ~/.config/spicetify/Themes/Turntable
-rm ~/.config/spicetify/Extensions/turntable.js
 ```
 
 2. config to spicetify default theme
@@ -74,11 +72,10 @@ rm ~/.config/spicetify/Extensions/turntable.js
 spicetify config current_theme SpicetifyDefault
 ```
 
-3. remove extension - Full App Display and Turntable(optional)
+1. remove extension - Full App Display
 
 ```shell
 spicetify config extensions fullAppDisplay.js-
-spicetify config extensions turntable.js-
 ```
 
 4. apply
